- Added script to delete all fleets not in their home galaxy to OSClient [Random Chaos] - Added "Scout Wave" method to Fleet Dialog to easily send one ship to many systems quickly [Random Chaos] - Added "Scrap Fleet" method in Fleet Dialog [Random Chaos] - Fixed IClientDB to handle __setitem__ commands where value==None (automatically was called during Scrap Fleet method) [Random Chaos] - Updated Starmap widget to allow appending a single fleet without full precompute [Random Chaos]
Note on Scout Wave: to prevent spam abuse, you can only select a system once per call instance. This can of course be easily overridden client side, but it keeps people who don't know how to program from sending 5000 1-ship fleets easily to a single target; though with scanner signature stacking, that's no longer as big an issue as it once was.

Note on Scrap Fleet/IClientDB: for some reason scrap fleet was calling IClientDB.__setitem__(key,None) [not sure what the key was]. In order to fix the call of None.__dict__, which is invalid, I wrapped the set in "if value != None:" which should catch any future instances of this error too. This is a client-side error, so it will not affect live server for anyone using the SVN version to scrap a fleet.

Note on deleting fleets not in home galaxy: This has been tested locally, but the test set is fairly small. Therefore, use with caution.

git-svn-id: svn://svn.code.sf.net/p/ospace/code/trunk@254 9ec6d687-074c-4f17-8882-0daec8a15dc2