Small directory restructuring such that client and server libs are contained under libs/ and the res/ and var/ folders are under shared. Applications when run from the source code will treat shared/ as their working directory. PyInstaller copies its contents to newly built distributions. Relative directory references that go go back up the directory tree (Ie. "../") created relative references like that in the finalized binary.
Remove unnecessary files from the repository such as the MANIFEST.in files and now-unused setup.py files.
Added specification files for the various apps to build with PyInstaller
Various adjustments in the existing programming to account for directory moves.
Add code for programming to detect if running in a bundle and setup folder relationships accordingly; Fix client.spec to point to the regular osc.py
Add metaskit script to server/ so that compat is easier to deal with.
Add var/db_data to .gitignore so we don't commit things we shouldn't be.

Fix segfault in PyGame caused by the tech tree generator because of a None reference being used unconditionally
Fix some server programming flat-out not working (opening var/token without catching IOException & splitting the RCS version and referencing a bad index)